Additional small changes
* Moved custom postgres functions into their own file
